+++ src/app/(app)/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <div>
+      <h1>contents</h1>
+    </div>
+  );
+}
