+++ src/app/layout.tsx
@@ -1,14 +1,7 @@
 import { Metadata } from "next";
-import "@/src/app/globals.css";
-import { Inter as FontSans } from "next/font/google";
+import "@/src/styles/tailwind.css";
 
 import React from "react";
-import { cn } from "@/src/lib/utils";
-
-const fontSans = FontSans({
-  subsets: ["latin"],
-  variable: "--font-sans",
-});
 
 export const metadata: Metadata = {
   title: "BeCreAIT",
@@ -23,14 +16,7 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <head />
-      <body
-        className={cn(
-          "min-h-screen bg-background font-sans antialiased",
-          fontSans.variable
-        )}
-      >
-        {children}
-      </body>
+      <body>{children}</body>
     </html>
   );
 }
