--- src/app/page.tsx
@@ -1,8 +1,0 @@
-export default function Home() {
-  return (
-    <>
-      <div>ss</div>
-      <p>ss</p>
-    </>
-  );
-}
