+++ src/styles/tailwind.css
@@ -1,0 +1,60 @@
+@import url("https://fonts.googleapis.com/css2?family=Offside&display=swap");
+
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+:root {
+  --foreground-rgb: 255, 255, 255;
+  --background-start-rgb: 0, 0, 0;
+  --background-end-rgb: 0, 0, 0;
+}
+
+body {
+  color: rgb(var(--foreground-rgb));
+  background: linear-gradient(
+    to bottom right,
+    rgb(var(--background-start-rgb)),
+    rgb(var(--background-end-rgb))
+  );
+}
+
+@layer utilities {
+  .text-balance {
+    text-wrap: balance;
+  }
+}
+
+.offside {
+  font-family: "Offside", cursive;
+}
+
+.neon_blue {
+  background: transparent;
+  color: #21cbff;
+  letter-spacing: 4px;
+  text-shadow:
+    0 0 5px #fff,
+    0 0 10px #fff,
+    0 0 20px #fff,
+    0 0 40px #0ff,
+    0 0 80px #0ff,
+    0 0 90px #0ff,
+    0 0 100px #0ff,
+    0 0 150px #0ff;
+}
+
+.neon_orange {
+  background: transparent;
+  color: #333;
+  letter-spacing: 4px;
+  text-shadow:
+    0 0 5px #fff,
+    0 0 10px #fff,
+    0 0 20px #fff,
+    0 0 40px #ee7800,
+    0 0 80px #ee7800,
+    0 0 90px #ee7800,
+    0 0 100px #ee7800,
+    0 0 150px #ee7800;
+}
